CleanUp

+++ AirGame/Client/Graphic/Renderers/AirBlowRender.cs
@@ -1,7 +1,6 @@
 ﻿using GlLib.Client.Api.Renderers;
 using GlLib.Client.Api.Sprites;
 using GlLib.Common.Entities;
-using GlLib.Utils;
 using GlLib.Utils.Math;
 
 namespace GlLib.Client.Graphic.Renderers
@@ -15,7 +14,7 @@ namespace GlLib.Client.Graphic.Renderers
             var layout = new TextureLayout(Textures.airBlow, 3, 1);
             _sprite = new LinearSprite(layout, 3, 10);
             var box = _e.AaBb;
-            _sprite.Scale((float) box.Width, (float) box.Height);
+            _sprite.Scale(box.Width, box.Height);
         }
 
         public override void Render(Entity _e, PlanarVector _xAxis, PlanarVector _yAxis)
