+++ AirGame/Client/Graphic/Renderers/CoinRenderer.cs
@@ -1,9 +1,7 @@
 using GlLib.Client.Api.Renderers;
 using GlLib.Client.Api.Sprites;
 using GlLib.Common.Entities;
-using GlLib.Utils;
 using GlLib.Utils.Math;
-using OpenTK;
 
 namespace GlLib.Client.Graphic.Renderers
 {
@@ -18,7 +16,7 @@ namespace GlLib.Client.Graphic.Renderers
 //            coinSprite.Translate(new PlanarVector(-8, -8));
 //            coinSprite.Scale(new Vector3(0.5f, 0.5f, 1));
             var box = _e.AaBb;
-            coinSprite.Scale((float) box.Width, (float) box.Height);
+            coinSprite.Scale(box.Width, box.Height);
         }
 
         public override void Render(Entity _e, PlanarVector _xAxis, PlanarVector _yAxis)
